refactored frontend to use React Context
